requirements.txt
allennlp == 0.9.0
ax-platform == 0.1.12
gensim == 3.8.3
plotly == 4.7.1
pytorch==1.5.1
spacy == 2.1.9
tensorboardx == 2.0
tokenizers == 0.7.0
numpy == 1.15.1
nltk == 3.5
python == 3.6.12
pytorch-pretrained-bert == 0.6.2
transformers == 2.9.1
